Move Contact, Tell Friends, Rate Us to Main Settings page
